fix: 관심 대분류를 단일 선택했을 경우, 다양성 보정 제외 (#20)

+++ src/main/java/com/chungbazi/server/domain/policy/application/dto/PolicyRecommendationContext.java
@@ -40,8 +40,8 @@ public record PolicyRecommendationContext(
         return interestSubCategories.contains(subCategory);
     }
 
-    public long interestCategoryCount(PolicyCategoryType category) {
-        return interestCategoryCounts.getOrDefault(category, 0L);
+    public int interestCategoryCount(PolicyCategoryType category) {
+        return interestCategoryCounts.getOrDefault(category, 0L).intValue();
     }
 
     public boolean hasLikedSubCategory(PolicySubCategoryType subCategory) {
